requirements.txt

colorama==0.4.6
fastapi==0.112.2
pydantic==2.8.2
python-dotenv==1.0.1
Requests==2.32.3
starlette==0.38.2
uvicorn==0.30.6
